requirements.txt
py7zr
mysql-connector-python
sqlalchemy
pandas
Unidecode
#python3-dev default-libmysqlclient-dev build-essential
mysqlclient
